requirements.txt
fastapi
uvicorn
python-dotenv
requests
langchain-openai
langgraph
pydantic